Integration-test specific conftest.py for Rocky
